+++ lib/lepton/src/Parser.cpp
@@ -178,7 +178,7 @@ ParsedExpression Parser::parse(const string& expression, const map<string, Custo
             vector<ParseToken> tokens = tokenize(subexpressions[i].substr(equalsPos+1));
             int pos = 0;
             subexpDefs[name] = parsePrecedence(tokens, pos, customFunctions, subexpDefs, 0);
-            if (pos != tokens.size())
+            if (pos != (int)tokens.size())
                 throw Exception("unexpected text at end of subexpression: "+tokens[pos].getText());
         }
 
@@ -187,7 +187,7 @@ ParsedExpression Parser::parse(const string& expression, const map<string, Custo
         vector<ParseToken> tokens = tokenize(primaryExpression);
         int pos = 0;
         ExpressionTreeNode result = parsePrecedence(tokens, pos, customFunctions, subexpDefs, 0);
-        if (pos != tokens.size())
+        if (pos != (int)tokens.size())
             throw Exception("unexpected text at end of expression: "+tokens[pos].getText());
         return ParsedExpression(result);
     }
@@ -198,7 +198,7 @@ ParsedExpression Parser::parse(const string& expression, const map<string, Custo
 
 ExpressionTreeNode Parser::parsePrecedence(const vector<ParseToken>& tokens, int& pos, const map<string, CustomFunction*>& customFunctions,
             const map<string, ExpressionTreeNode>& subexpressionDefs, int precedence) {
-    if (pos == tokens.size())
+  if (pos == (int)tokens.size())
         throw Exception("unexpected end of expression");
 
     // Parse the next value (number, variable, function, parenthesized expression)
@@ -224,7 +224,7 @@ ExpressionTreeNode Parser::parsePrecedence(const vector<ParseToken>& tokens, int
     else if (token.getType() == ParseToken::LeftParen) {
         pos++;
         result = parsePrecedence(tokens, pos, customFunctions, subexpressionDefs, 0);
-        if (pos == tokens.size() || tokens[pos].getType() != ParseToken::RightParen)
+        if (pos == (int)tokens.size() || tokens[pos].getType() != ParseToken::RightParen)
             throw Exception("unbalanced parentheses");
         pos++;
     }
@@ -238,7 +238,7 @@ ExpressionTreeNode Parser::parsePrecedence(const vector<ParseToken>& tokens, int
             if (moreArgs)
                 pos++;
         } while (moreArgs);
-        if (pos == tokens.size() || tokens[pos].getType() != ParseToken::RightParen)
+        if (pos == (int)tokens.size() || tokens[pos].getType() != ParseToken::RightParen)
             throw Exception("unbalanced parentheses");
         pos++;
         Operation* op = getFunctionOperation(token.getText(), customFunctions);
